requirements.txt
langchain-google-genai
python-dotenv
langchain
langchain-community
unstructured[pdf]
faiss-cpu
streamlit
streamlit-chat